Revert some unused changes and cleaned up some warnings.
