+++ build.gradle
@@ -16,6 +16,8 @@
 
 
 
+
+
 allprojects {
     apply plugin: 'idea'
     apply plugin: 'eclipse'
@@ -45,6 +47,10 @@ subprojects {
         testCompile group: 'org.hamcrest', name: 'hamcrest-all', version: '1.+'
         testCompile 'org.testng:testng:6.8.+'
         testCompile 'org.mockito:mockito-all:1.10.+'
+    }
+
+    tasks.withType(JavaCompile) {
+        options.compilerArgs << "-Xlint:unchecked"
     }
 
     //noinspection GroovyAssignabilityCheck
