+++ populace-core/src/main/java/org/datalorax/populace/core/populate/instance/BigDecimalInstanceFactory.java
@@ -43,10 +43,10 @@ public class BigDecimalInstanceFactory implements InstanceFactory {
         this.defaultValue = defaultValue;
     }
 
+    @SuppressWarnings("unchecked")
     @Override
     public <T> T createInstance(Class<? extends T> rawType, final Object parent, final InstanceFactories instanceFactories) {
         Validate.isTrue(rawType.equals(BigDecimal.class), "BigDecimal type expected. Got: %s", rawType);
-        //noinspection unchecked
         return (T) defaultValue;
     }
